feat(node-drivers): split mysql and PlanetScale implementations
