+++ packages/client/src/runtime/core/engines/library/LibraryEngine.ts
@@ -39,7 +39,8 @@ import { getBatchRequestPayload } from '../common/utils/getBatchRequestPayload'
 import { getErrorMessageWithLink } from '../common/utils/getErrorMessageWithLink'
 import { getInteractiveTransactionId } from '../common/utils/getInteractiveTransactionId'
 import { DefaultLibraryLoader } from './DefaultLibraryLoader'
-import { createMySQLDriver, createPlanetscaleDriver } from './driver/mysql'
+import { createMySQLDriver } from './driver/mysql'
+import { createPlanetscaleDriver } from './driver/planetScale'
 import { type BeforeExitListener, ExitHooks } from './ExitHooks'
 import type { Library, LibraryLoader, QueryEngineConstructor, QueryEngineInstance } from './types/Library'
