+++ packages/client/src/runtime/core/engines/library/driver/mysql.ts
@@ -1,9 +1,9 @@
 /* eslint-disable */
 
 import * as mysql from 'mysql2/promise'
-import * as planetScale from '@planetscale/database' 
 import type { Driver, Closeable, ResultSet, Query } from '../types/Library'
 import { ColumnType } from '../types/Library'
+import { binder } from './utils/binder'
 
 /**
  * This is a simplification of quaint's value inference logic. Take a look at quaint's conversion.rs
@@ -62,91 +62,6 @@ function fieldToColumnType({ type: typeIdx }: mysql.FieldPacket): ColumnType {
   }
   
   return colType
-}
-
-function planetScaleFieldToColumnType(type: 'INT32' | 'VARCHAR' | string): ColumnType {
-  switch (type) {
-    case 'INT32':
-      return ColumnType.Int64 // trick Quaint into thinking it's a 64-bit integer
-    case 'VARCHAR':
-      return ColumnType.Text
-    default:
-      throw new Error(`Unsupported planetscale type: ${type}`)
-  }
-}
-
-type PlanetscaleConfig = {
-  host: string,
-  username: string,
-  password: string,
-}
-
-class PrismaPlanetscale implements Driver, Closeable {
-  private client: planetScale.Connection
-  private maybeVersion?: string
-  private isRunning: boolean = true
-
-  constructor(config: PlanetscaleConfig) {
-    this.client = planetScale.connect(config)
-
-    // lazily retrieve the version and store it into `maybeVersion`
-    this.client.execute('SELECT @@version, @@GLOBAL.version').then((results) => {
-      this.maybeVersion = results.rows[0]['@@version']
-    })
-  }
-
-  async close(): Promise<void> {
-    if (this.isRunning) {
-      this.isRunning = false
-    }
-  }
-
-  /**
-   * Returns false, if connection is considered to not be in a working state.
-   */
-  isHealthy(): boolean {
-    const result = this.maybeVersion !== undefined
-      && this.isRunning
-    return result
-  }
-  
-  /**
-   * Execute a query given as SQL, interpolating the given parameters.
-   */
-  async queryRaw(query: Query): Promise<ResultSet> {
-    const { sql, args: values } = query
-    const { fields, rows: results } = await this.client.execute(sql, values, { as: 'object' })
-
-    const columns = fields.map(field => field.name)
-    const resultSet: ResultSet = {
-      columnNames: columns,
-      columnTypes: fields.map(field => planetScaleFieldToColumnType(field.type)),
-      rows: results.map(result => columns.map(column => result[column])),
-    }
-
-    return resultSet
-  }
-
-  /**
-   * Execute a query given as SQL, interpolating the given parameters and
-   * returning the number of affected rows.
-   * Note: Queryable expects a u64, but napi.rs only supports u32.
-   */
-  async executeRaw(query: Query): Promise<number> {
-    const { sql, args: values } = query
-    const { rowsAffected } = await this.client.execute(sql, values)
-    return rowsAffected
-  }
-
-  /**
-   * Return the version of the underlying database, queried directly from the
-   * source. This corresponds to the `version()` function on PostgreSQL for
-   * example. The version string is returned directly without any form of
-   * parsing or normalization.
-   */
-  version(): Promise<string | undefined> {
-    return Promise.resolve(this.maybeVersion)
-  }
 }
 
 class PrismaMySQL implements Driver, Closeable {
@@ -227,20 +142,7 @@ class PrismaMySQL implements Driver, Closeable {
   }
 }
 
-const binder = (driver: Driver & Closeable): Driver & Closeable => ({
-  queryRaw: driver.queryRaw.bind(driver),
-  executeRaw: driver.executeRaw.bind(driver),
-  version: driver.version.bind(driver),
-  isHealthy: driver.isHealthy.bind(driver),
-  close: driver.close.bind(driver),
-})
-
 export const createMySQLDriver = (connectionString: string): Driver & Closeable => {
   const db = new PrismaMySQL(connectionString)
-  return binder(db)
-}
-
-export const createPlanetscaleDriver = (config: PlanetscaleConfig): Driver & Closeable => {
-  const db = new PrismaPlanetscale(config)
   return binder(db)
 }
